apu: Move MCPX APU in separate directory
