+++ hw/xbox/Makefile.objs
@@ -3,13 +3,13 @@ obj-y += chihiro.o
 obj-y += xbox_pci.o acpi_xbox.o
 obj-y += amd_smbus.o smbus_xbox_smc.o smbus_cx25871.o smbus_adm1032.o
 obj-y += nvnet.o
-obj-y += mcpx_apu.o mcpx_aci.o
+obj-y += mcpx_aci.o
 obj-y += lpc47m157.o
 obj-y += xid.o xid-sdl.o
 obj-y += chihiro-usb.o
 
-obj-y += dsp/
 obj-y += nv2a/
+obj-y += mcpx_apu/
 
 xid-sdl.o-cflags := $(SDL_CFLAGS)
 xid-sdl.o-libs := $(SDL_LIBS)
