+++ hw/xbox/mcpx_apu/mcpx_apu_debug.h
